backend/requirements.txt
fastapi==0.115.0
starlette==0.38.6
uvicorn[standard]==0.30.6
sqlalchemy==2.0.48
alembic==1.17.0
pydantic-settings==2.6.1
python-jose==3.5.0
passlib==1.7.4
python-multipart==0.0.12
requests==2.33.0
email-validator==2.3.0
